+++ xiaohe66-web-portal/src/main/java/com/xiaohe66/web/code/resume/dao/ResumeProjectDao.java
@@ -17,5 +17,5 @@ public interface ResumeProjectDao extends BaseDao<ResumeProject>{
      * @param resumeId 简历id
      * @return List<ResumeProject>
      */
-    List<ResumeProject> findByResumeId(@Param("resumeId") Long resumeId);
+    List<ResumeProject> findByResumeId(@Param("resumeId") Integer resumeId);
 }
